third_party/blink/web_tests/platform/mac-mac12/external/wpt/service-workers/service-worker/tentative/static-router/static-router-resource-timing.https-expected.txt

This is a testharness.js-based test.
[FAIL] Main resource matched the rule with fetch-event source
  assert_equals: fetch-event as source on main resource expected (string) "fetch-event" but got (undefined) undefined
[FAIL] Main resource load matched with the condition and resource timing
  assert_equals: network as source on main resource expected (string) "network" but got (undefined) undefined
[FAIL] Main resource load not matched with the condition and resource timing
  assert_equals: no rule matched on main resource expected (string) "" but got (undefined) undefined
[FAIL] Main resource load matched with the cache source and resource timing
  assert_equals: cache as source on main resource and cache hit expected (string) "cache" but got (undefined) undefined
[FAIL] Main resource fallback to the network when there is no cache entry and resource timing
  assert_equals: cache as source on main resource and cache miss, fallback to network expected (string) "cache" but got (undefined) undefined
[FAIL] Subresource load matched the rule fetch-event source
  assert_equals: fetch-event as source on sub resource expected (string) "fetch-event" but got (undefined) undefined
[FAIL] Subresource load not matched with URLPattern condition
  assert_equals: no source type matched expected (string) "" but got (undefined) undefined
[FAIL] Subresource load matched with URLPattern condition
  assert_equals: network as source on subresource expected (string) "network" but got (undefined) undefined
[FAIL] Subresource load matched with the cache source rule
  assert_equals: cache as source on subresource and cache hits expected (string) "cache" but got (undefined) undefined
[FAIL] Subresource load did not match with the cache and fallback to the network
  assert_equals: cache as source on subresource and cache misses expected (string) "cache" but got (undefined) undefined
[FAIL] Main resource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response
  assert_equals: race as source on main resource, and fetch-event wins expected (string) "race-network-and-fetch" but got (undefined) undefined
[FAIL] Main resource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler
  assert_equals: race as source on main resource, and network wins expected (string) "race-network-and-fetch" but got (undefined) undefined
[FAIL] Subresource load matched the rule with race-network-and-fetch-handler source, and the fetch handler response is faster than the server response
  assert_equals: race as source on subresource and fetch wins expected (string) "race-network-and-fetch" but got (undefined) undefined
[FAIL] Subresource load matched the rule with race-network-and-fetch-handler source, and the server reseponse is faster than the fetch handler
  assert_equals: race as source on subresource and network wins expected (string) "race-network-and-fetch" but got (undefined) undefined
Harness: the test ran to completion.
